22.04.2023

+++ lib/current_mix/current_mix.dart
@@ -21,7 +21,7 @@ class InitalScreenWidget extends StatefulWidget {
 }
 
 class HomeWidget extends State<InitalScreenWidget> {
-  final int _counter = context.read<DataTypeList>().activePlayList.length;
+  final int _counter = context.read<DataTypeList>().activePlayList.length.toInt();
 
   void _incrementCounter() {
     setState(() {
